Add log forwarding using filebeat
Forwarding to an external ELK stack is working. I need to refactor
the filebeat client into a role on it's own eventually, but I'm
not sure the best way to proceed with testing.
In my setup, I used a [docker based elk
stack](https://github.com/blacktop/docker-elk) and logs were coming
through. There is still a fair bit of work that needs to be done
to make these logs useful and searchable, as at the moment ES
does not know how to parse the fields in our logs.
Sending logs over SSL will require some orchestration of certs,
but manually testing it appeared to work.